+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Always destroy old access token after successful Hub API call in `Zaikio::AccessToken#refresh!`
+
 ## 0.7.2 - 2021-03-25
 
 * Replace dependency on `rails` with a more specific dependency on `railties` and friends